requirements.txt
google-generativeai
sounddevice
scipy
python-dotenv
numpy
soundfile
faster_whisper
pytz
